# ui/web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Strategy Lab",
  "version": "1"
 },
 "paths": {
...
  "/api/research/result": {
   "get": {
    "summary": "Result",
    "operationId": "result_api_research_result_get",
    "parameters": [
     {
      "name": "key",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResultView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/research/curves": {
   "get": {
    "summary": "Curves",
    "operationId": "curves_api_research_curves_get",
    "parameters": [
     {
      "name": "keys",
      "in": "query",
      "required": true,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "maxItems": 16,
       "title": "Keys"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Curves"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/runs": {
   "post": {
    "summary": "Run Start",
    "operationId": "run_start_api_runs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunStart"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunAccepted"
        }
       }
      }
     },
     "409": {
      "description": "Conflict",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Refusal"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/runs/{run_id}/stop": {
   "post": {
    "summary": "Run Stop",
    "operationId": "run_stop_api_runs__run_id__stop_post",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunAccepted"
        }
       }
      }
     },
     "404": {
      "description": "Not Found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Refusal"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/runs/{run_id}/resume": {
   "post": {
    "summary": "Run Resume",
    "operationId": "run_resume_api_runs__run_id__resume_post",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunResume"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunAccepted"
        }
       }
      }
     },
     "404": {
      "description": "Not Found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Refusal"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AssetRow": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "strategy": {
      "type": "string",
      "title": "Strategy"
     },
     "list_id": {
      "type": "string",
      "title": "List Id"
     },
     "timeframe": {
      "type": "string",
      "enum": [
       "1h",
       "4h",
       "1d"
      ],
      "title": "Timeframe"
     },
     "start": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start"
     },
     "end": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "End"
     },
     "months": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Months"
     },
     "avg_monthly": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Monthly"
     },
     "cagr": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cagr"
     },
     "green": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Green"
     },
     "max_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Dd"
     },
     "at_target_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd"
     },
     "at_target_dd_multiple": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd Multiple"
     },
     "at_target_dd_5y": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd 5Y"
     },
     "sharpe": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sharpe"
     },
     "k_ratio": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "K Ratio"
     },
     "trades": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trades"
     },
     "beats_bh": {
      "type": "boolean",
      "title": "Beats Bh"
     },
     "vs_bh": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vs Bh"
     },
     "bh_cash": {
      "type": "boolean",
      "title": "Bh Cash"
     },
     "bh_sharpe": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh Sharpe"
     },
     "bh_max_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh Max Dd"
     },
     "bh_k_ratio": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh K Ratio"
     },
     "bh_at_target_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh At Target Dd"
     },
     "bh_at_target_dd_multiple": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh At Target Dd Multiple"
     },
     "targets": {
      "items": {
       "type": "boolean"
      },
      "type": "array",
      "title": "Targets"
     },
     "targets_met": {
      "type": "integer",
      "title": "Targets Met"
     },
     "robustness": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Robustness"
       },
       {
        "type": "null"
       }
      ]
     },
     "loses": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Loses"
     },
     "stale": {
      "type": "boolean",
      "title": "Stale"
     },
     "instrument_id": {
      "type": "string",
      "title": "Instrument Id"
     },
     "bh_avg_monthly": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh Avg Monthly"
     },
     "params_now": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params Now"
     },
     "held": {
      "type": "boolean",
      "title": "Held"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "key",
     "id",
     "strategy",
     "list_id",
     "timeframe",
     "start",
     "end",
     "months",
     "avg_monthly",
     "cagr",
     "green",
     "max_dd",
     "at_target_dd",
     "at_target_dd_multiple",
     "at_target_dd_5y",
     "sharpe",
     "k_ratio",
     "trades",
     "beats_bh",
     "vs_bh",
     "bh_cash",
     "bh_sharpe",
     "bh_max_dd",
     "bh_k_ratio",
     "bh_at_target_dd",
     "bh_at_target_dd_multiple",
     "targets",
     "targets_met",
     "robustness",
     "loses",
     "stale",
     "instrument_id",
     "bh_avg_monthly",
     "params_now",
     "held"
    ],
    "title": "AssetRow"
   },
...
   "BenchmarkView": {
    "properties": {
     "pts": {
      "items": {
       "prefixItems": [
        {
         "type": "integer"
        },
        {
         "type": "number"
        }
       ],
       "type": "array",
       "maxItems": 2,
       "minItems": 2
      },
      "type": "array",
      "title": "Pts"
     },
     "months": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Months"
       },
       {
        "type": "null"
       }
      ]
     },
     "growth": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Growth"
     },
     "figures": {
      "$ref": "#/components/schemas/BenchmarkFigures"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "pts",
     "months",
     "growth",
     "figures"
    ],
    "title": "BenchmarkView"
   },
   "Candidate": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "strategy": {
      "type": "string",
      "title": "Strategy"
     },
     "sharpe": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sharpe"
     },
     "windows": {
      "type": "integer",
      "title": "Windows"
     },
     "stale": {
      "type": "boolean",
      "title": "Stale"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "key",
     "strategy",
     "sharpe",
     "windows",
     "stale"
    ],
    "title": "Candidate"
   },
...
   "CurveView": {
    "properties": {
     "pts": {
      "items": {
       "prefixItems": [
        {
         "type": "integer"
        },
        {
         "type": "number"
        }
       ],
       "type": "array",
       "maxItems": 2,
       "minItems": 2
      },
      "type": "array",
      "title": "Pts"
     },
     "dd": {
      "items": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "array",
      "title": "Dd"
     },
     "months": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Months"
       },
       {
        "type": "null"
       }
      ]
     },
     "growth": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Growth"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "pts",
     "dd",
     "months",
     "growth"
    ],
    "title": "CurveView"
   },
   "Curves": {
    "properties": {
     "curves": {
      "additionalProperties": {
       "$ref": "#/components/schemas/CurveLine"
      },
      "type": "object",
      "title": "Curves"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "curves"
    ],
    "title": "Curves"
   },
   "Elsewhere": {
    "properties": {
     "lists": {
      "items": {
       "$ref": "#/components/schemas/Cell"
      },
      "type": "array",
      "title": "Lists"
     },
     "alone": {
      "items": {
       "$ref": "#/components/schemas/Cell"
      },
      "type": "array",
      "title": "Alone"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "lists",
     "alone"
    ],
    "title": "Elsewhere"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "IndexView": {
    "properties": {
     "pts": {
      "items": {
       "prefixItems": [
        {
         "type": "integer"
        },
        {
         "type": "number"
        }
       ],
       "type": "array",
       "maxItems": 2,
       "minItems": 2
      },
      "type": "array",
      "title": "Pts"
     },
     "months": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Months"
       },
       {
        "type": "null"
       }
      ]
     },
     "growth": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Growth"
     },
     "figures": {
      "$ref": "#/components/schemas/BenchmarkFigures"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "pts",
     "months",
     "growth",
     "figures",
     "name"
    ],
    "title": "IndexView"
   },
...
   "ListRow": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "strategy": {
      "type": "string",
      "title": "Strategy"
     },
     "list_id": {
      "type": "string",
      "title": "List Id"
     },
     "timeframe": {
      "type": "string",
      "enum": [
       "1h",
       "4h",
       "1d"
      ],
      "title": "Timeframe"
     },
     "start": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start"
     },
     "end": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "End"
     },
     "months": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Months"
     },
     "avg_monthly": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Monthly"
     },
     "cagr": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cagr"
     },
     "green": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Green"
     },
     "max_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Dd"
     },
     "at_target_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd"
     },
     "at_target_dd_multiple": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd Multiple"
     },
     "at_target_dd_5y": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "At Target Dd 5Y"
     },
     "sharpe": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sharpe"
     },
     "k_ratio": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "K Ratio"
     },
     "trades": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trades"
     },
     "beats_bh": {
      "type": "boolean",
      "title": "Beats Bh"
     },
     "vs_bh": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vs Bh"
     },
     "bh_cash": {
      "type": "boolean",
      "title": "Bh Cash"
     },
     "bh_sharpe": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh Sharpe"
     },
     "bh_max_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh Max Dd"
     },
     "bh_k_ratio": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh K Ratio"
     },
     "bh_at_target_dd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh At Target Dd"
     },
     "bh_at_target_dd_multiple": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bh At Target Dd Multiple"
     },
     "targets": {
      "items": {
       "type": "boolean"
      },
      "type": "array",
      "title": "Targets"
     },
     "targets_met": {
      "type": "integer",
      "title": "Targets Met"
     },
     "robustness": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Robustness"
       },
       {
        "type": "null"
       }
      ]
     },
     "loses": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Loses"
     },
     "stale": {
      "type": "boolean",
      "title": "Stale"
     },
     "worst_month": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Worst Month"
     },
     "evaluated_at": {
      "type": "string",
      "title": "Evaluated At"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "key",
     "id",
     "strategy",
     "list_id",
     "timeframe",
     "start",
     "end",
     "months",
     "avg_monthly",
     "cagr",
     "green",
     "max_dd",
     "at_target_dd",
     "at_target_dd_multiple",
     "at_target_dd_5y",
     "sharpe",
     "k_ratio",
     "trades",
     "beats_bh",
     "vs_bh",
     "bh_cash",
     "bh_sharpe",
     "bh_max_dd",
     "bh_k_ratio",
     "bh_at_target_dd",
     "bh_at_target_dd_multiple",
     "targets",
     "targets_met",
     "robustness",
     "loses",
     "stale",
     "worst_month",
     "evaluated_at"
    ],
    "title": "ListRow"
   },
...
   "NameView": {
    "properties": {
     "instrument_id": {
      "type": "string",
      "title": "Instrument Id"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "trades": {
      "type": "integer",
      "title": "Trades"
     },
     "won": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Won"
     },
     "compounded": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compounded"
     },
     "alone": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Alone"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "instrument_id",
     "label",
     "trades",
     "won",
     "compounded",
     "alone"
    ],
    "title": "NameView"
   },
   "Refusal": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     },
     "confirm": {
      "type": "boolean",
      "title": "Confirm"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "message",
     "confirm"
    ],
    "title": "Refusal"
   },
   "ResultView": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "kind": {
      "type": "string",
      "enum": [
       "list",
       "asset",
       "pick"
      ],
      "title": "Kind"
     },
     "row": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ListRow"
       },
       {
        "$ref": "#/components/schemas/AssetRow"
       }
      ],
      "title": "Row"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notes"
     },
     "fill": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fill"
     },
     "evaluated_at": {
      "type": "string",
      "title": "Evaluated At"
     },
     "figures": {
      "$ref": "#/components/schemas/Figures"
     },
     "curve": {
      "$ref": "#/components/schemas/CurveView"
     },
     "benchmark": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BenchmarkView"
       },
       {
        "type": "null"
       }
      ]
     },
     "index": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/IndexView"
       },
       {
        "type": "null"
       }
      ]
     },
     "windows": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/WindowsView"
       },
       {
        "type": "null"
       }
      ]
     },
     "params": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params"
     },
     "grid": {
      "anyOf": [
       {
        "additionalProperties": {
         "items": {},
         "type": "array"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grid"
     },
     "params_now": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params Now"
     },
     "names": {
      "items": {
       "$ref": "#/components/schemas/NameView"
      },
      "type": "array",
      "title": "Names"
     },
     "elsewhere": {
      "$ref": "#/components/schemas/Elsewhere"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/Candidate"
      },
      "type": "array",
      "title": "Candidates"
     },
     "beyond_luck": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Beyond Luck"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "key",
     "kind",
     "row",
     "description",
     "notes",
     "fill",
     "evaluated_at",
     "figures",
     "curve",
     "benchmark",
     "index",
     "windows",
     "params",
     "grid",
     "params_now",
     "names",
     "elsewhere",
     "candidates",
     "beyond_luck"
    ],
    "title": "ResultView"
   },
...
   "RunAccepted": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id"
    ],
    "title": "RunAccepted"
   },
...
   "RunResume": {
    "properties": {
     "confirm": {
      "type": "boolean",
      "title": "Confirm",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "RunResume"
   },
   "RunStart": {
    "properties": {
     "single_assets": {
      "type": "boolean",
      "title": "Single Assets",
      "default": false
     },
     "everything": {
      "type": "boolean",
      "title": "Everything",
      "default": false
     },
     "only": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunOnly"
       },
       {
        "type": "null"
       }
      ]
     },
     "confirm": {
      "type": "boolean",
      "title": "Confirm",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "RunStart"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WindowsView": {
    "properties": {
     "sets": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Sets"
     },
     "end": {
      "type": "string",
      "title": "End"
     },
     "train": {
      "type": "integer",
      "title": "Train"
     },
     "w": {
      "items": {
       "prefixItems": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ],
       "type": "array",
       "maxItems": 2,
       "minItems": 2
      },
      "type": "array",
      "title": "W"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "sets",
     "end",
     "train",
     "w"
    ],
    "title": "WindowsView"
   }
  }
 }
}